+++ config/zones/SV.yaml
@@ -206,6 +206,6 @@ fallbackZoneMixes:
         unknown: 0.057362415139040494
         wind: 0.03768562954230214
 parsers:
-  production: SV.fetch_production
+  production: estadistico.fetch_production
   productionCapacity: EMBER.fetch_production_capacity
 timezone: America/El_Salvador
